pyproject.toml
[project]
name = "ianews"
version = "0.1.0"
description = "Agrega noticias sobre IA desde RSS/Atom y las guarda en SQLite."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.11",
    "httpx>=0.27.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
]

[project.scripts]
ianews = "ianews.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
